==> 1_Build_and_clean_dataset_github_Feb_14_2025.Rmd <==
---
title: "Data cleaning"
date: "`r Sys.Date()`"
output: html_document
---

Purpose: create a dataset of NHANES cycles 1, 7, and 8 (1999-2000, 2011-2012, and 2013-2014) containing the variables needed for the cognition ExWAS analyses; clean the chemical exposure dataset

Participant inclusion criteria: in cycles 1, 7, or 8 (cycles with both cognitive measures and fish consumption measures), aged 60+, with at least one chemical exposure measurement

Data sources: 

https://www.kaggle.com/datasets/nguyenvy/nhanes-19882018

https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(haven)
library(sjlabelled)
```

# Load and prepare NHANES files
```{r}
# load source environment
load('Data/w - nhanes_1988_2018.RData')

# remove files not needed
rm(list = setdiff(ls(), c('chemicals_clean', 'comments_clean', 'demographics_clean', 'dietary_clean',
                          'occupation_clean', 'questionnaire_clean', 'response_clean', 'weights_clean')))

# select variables needed from demographics, occupation, questionnaire, and response files
demographics_clean <- demographics_clean %>%
  select(SEQN, SEQN_new, SDDSRVYR, RIDAGEYR, RIAGENDR, RIDRETH1, INDFMPIR, DMDEDUC2, DMDEDUC3,
         SDMVPSU, SDMVSTRA, WTMEC2YR, WTMEC4YR, WTINT2YR, WTINT4YR)

occupation_clean <- occupation_clean %>%
  select(SEQN, SEQN_new, SDDSRVYR, OCD150, OCQ380)

questionnaire_clean <- questionnaire_clean %>%
  select(SEQN, SEQN_new, SDDSRVYR, ALQ101, ALQ120Q, ALQ120U)
  
response_clean <- response_clean %>%
  select(SEQN, SEQN_new, SDDSRVYR, URXUCR, BMXBMI, BMXWAIST, VNEGFR)

# merge
nhanes_data <- left_join(demographics_clean,
                         occupation_clean,
                         by = c('SEQN', 'SEQN_new', 'SDDSRVYR')) %>%
    left_join(.,
            questionnaire_clean,
            by = c('SEQN', 'SEQN_new', 'SDDSRVYR')) %>%
  left_join(.,
            response_clean,
            by = c('SEQN', 'SEQN_new', 'SDDSRVYR'))
dim(nhanes_data) # 135310     24

# filter to cycles 1, 7, 8 (years 1999-2000, 2011-2012, 2013-2014)
nhanes_data <- nhanes_data %>%
  filter(SDDSRVYR %in% c(1, 7, 8))
dim(nhanes_data) # 29896    24

# create an initial indicator for adults 60+
nhanes_data <- nhanes_data %>%
  mutate(aged_60plus = case_when(RIDAGEYR >= 60 ~ 1,
                                 .default = 0))

table(nhanes_data$aged_60plus, useNA = 'always')
# N = 5466

# pull SEQN identifiers for these participants
aged_60plus <- nhanes_data %>%
  filter(aged_60plus == 1) %>%
  pull(SEQN)

# load data dictionary and filter to chemicals
chem_dictionary <- read_csv('Data/dictionary_nhanes.csv') %>%
  filter(in_dataset == 'Chemicals') %>%
  select(variable_codename_use, variable_description_use, unique_cycles,
         comment_codename_use, chemical_family)

# why 598 variables in chemicals_clean but 599 entries in dictionary?
#chem_find_extra <- chem_dictionary %>%
#  filter(!variable_codename_use %in% colnames(chemicals_clean))
# nothing missing - checked dictionary manually and there is a duplicate entry for LBXTC2
# remove duplicate
chem_dictionary <- chem_dictionary %>%
  distinct()

```

# Chemical and participant inclusion/exclusion

Chemical exclusion criteria: 

 - start with chemicals measured in the relevant cycles (1999-2000, 2011-2012, 2013-2014) --> 395
 
 - drop some chemical families --> 366
 
 - drop lipid-unadjusted variables when an adjusted version exists --> 316
 
 - drop chemicals if over 50% of measurements are below lower limit of detection --> 147
 
## Filter chemicals dataset to adults 60+ in relevant cycles
```{r}
chems <- chemicals_clean %>%
  filter(SDDSRVYR %in% c(1, 7, 8)) %>%
  filter(SEQN %in% aged_60plus)
dim(chems) # 5078  598

# note that aged_60plus is 5466 participants while chems contains 5078: not all participants have chemical measures so this is fine

```

## Preliminary exclusion: drop any chems not measured in the relevant cycles
```{r}
# identify chems measured in cycles 1, 7, 8
chem_dictionary_3cycles <- chem_dictionary %>%
  filter(grepl(paste(c('1999-2000', '2011-2012', '2013-2014'), collapse = '|'), unique_cycles))
dim(chem_dictionary_3cycles) # 398   5 --> 3 rows are for identifiers, so 395 chems were measured in the cycles being used

# reduce chems dataset to these variables
chems <- chems %>%
  select(chem_dictionary_3cycles$variable_codename_use)
dim(chems) # 5078  398 --> 3 identifiers, 395 chems

```

## Drop dietary components, phytoestrogens, and nicotine metabolites measured only in smokers
```{r}
nicotine_to_remove <- c('URXNOXT', 'URXCOXT', 'URXNNCT', 'URXANTT', 'URXANBT', 'URXNICT')
# see https://wwwn.cdc.gov/Nchs/Nhanes/2013-2014/UCOT_H.htm: 
# these chems were only measured if urinary cotinine > 20 ng/mL (as indicator of smoking)

dictionary_included <- chem_dictionary_3cycles %>%
  filter(!chemical_family %in% c('Dietary Components',
                                 'Phytoestrogens')) %>%
  filter(!variable_codename_use %in% nicotine_to_remove)
dim(dictionary_included) # 369   5

chems <- chems %>%
  select(dictionary_included$variable_codename_use)
dim(chems) # 5078  369

```

## Drop lipid-unadjusted chems
```{r}
# some vars have 2 versions: unadjusted and lipid-adjusted versions - want to use the adjusted ones
# use data dictionary to first identify the lipid-adjusted chems using the descriptions (variable_description_use)
# then identify the unadjusted versions to drop

# pull info for lipid-adjusted chems
lipid_adj_dictionary <- dictionary_included %>%
  filter(grepl('Lipid Adj', variable_description_use))

# make a separate variable of the chem names, dropping the 'LA' to get the unadjusted versions
# example: LBX028LA -> LBX028
lipid_adj_dictionary <- lipid_adj_dictionary %>%
  mutate(chem_codenames_unadj = gsub('LA$', '', variable_codename_use))

# update dictionary and chems dataset to remove the lipid unadj variables
dictionary_included <- dictionary_included %>%
  filter(!variable_codename_use %in% lipid_adj_dictionary$chem_codenames_unadj)
dim(dictionary_included) # 319   5

chems <- chems %>%
  select(dictionary_included$variable_codename_use)
dim(chems) # 5078  319

```

## Drop chems with over 50% of measurements below limit of detection (LOD)
```{r}
# filter comments dataset to participants aged 60+ in relevant cycles
chem_comments <- comments_clean %>% 
  filter(SDDSRVYR %in% c(1, 7, 8)) %>%
  filter(SEQN %in% aged_60plus)
dim(chem_comments) # 5078  505

# keep only the comment vars for currently included chems
comment_names <- dictionary_included %>%
  drop_na(comment_codename_use) %>%
  pull(comment_codename_use)

chem_comments <- chem_comments %>%
  select(SEQN, SEQN_new, SDDSRVYR, all_of(comment_names))
dim(chem_comments) # 5078  319

# process for determining % above LOD
# sum of count above (comment = 0)
# sum of count below (comment = 1)
# above + below = total participants with measurement
# above / total * 100 = % above LOD

# turn comments into a long dataset: one chem measurement/comment per row
long_comments <- chem_comments %>%
  pivot_longer(cols = all_of(comment_names),
               values_to = 'comment',
               names_to = 'comment_codename') %>%
  drop_na(comment) %>% # if the comment is NA, then that measurement doesn't exist for that participant - drop
  select(comment_codename, comment)
dim(long_comments) # 285518      2

# calculate number of measurements above vs below LOD for each chem
lod_calc <- long_comments %>%
  group_by(comment_codename) %>%
  summarize(n_above_lod = sum(comment == 0), 
            n_below_lod = sum(comment == 1),
            n_measures = n_above_lod + n_below_lod,
            pct_above_lod = (n_above_lod / n_measures) * 100) %>%
  mutate(use_chem = ifelse(pct_above_lod >= 50, 1, 0))

table(lod_calc$use_chem, useNA = 'always')

# look at just the chems to keep
above_50_pct <- lod_calc %>%
  filter(use_chem == 1)

# use the dictionary to match which comment codename goes to which chemical and filter it to just the chems to keep
dictionary_included <- dictionary_included %>%
  filter(comment_codename_use %in% above_50_pct$comment_codename)
dim(dictionary_included) # 147   5
# identifiers dropped off of dictionary here, so 147 chems now

# keep only these vars in the chems dataset
chems <- chems %>%
  select(SEQN, SEQN_new, SDDSRVYR, all_of(dictionary_included$variable_codename_use))
dim(chems) # 5078  150 --> 3 identifiers, 147 chems

```

## Drop participants without any chem measures
```{r}
# starting N = 5078 (chems dataset)
# create long dataset, one row per measure, in order to drop off all NAs together
drop_chem_nas <- chems %>%
  pivot_longer(cols = all_of(dictionary_included$variable_codename_use),
               names_to = 'chem',
               values_to = 'measure') %>%
  drop_na(measure)
dim(drop_chem_nas) # 147790      5

# pull sample size of participants with at least one measure
print(length(unique(drop_chem_nas$SEQN)))
# N = 5078 -> N = 4982

# keep only these participants in chems dataset
chems <- chems %>%
  filter(SEQN %in% unique(drop_chem_nas$SEQN))
dim(chems) # 4982  150

```

## Update analysis inclusion variable based on chemical measure availability
```{r}
# nhanes_data$aged_60plus N = 5466
# chems participant N = 4982 (additionally dropped anyone without chem data)

included_seqns <- chems$SEQN

nhanes_data <- nhanes_data %>%
  mutate(include = case_when(SEQN %in% included_seqns ~ 1,
                             .default = 0))

table(nhanes_data$include, useNA = 'always')
table(nhanes_data$aged_60plus, nhanes_data$include, useNA = 'always')
# confirmed N = 4982 to include

```

## Add serum cotinine to nhanes_sample
```{r}
cotinine <- chemicals_clean %>%
  filter(SDDSRVYR %in% c(1, 7, 8)) %>% 
  select(SEQN, SDDSRVYR, LBXCOT)

nhanes_data <- left_join(nhanes_data,
                         cotinine,
                         by = c('SEQN', 'SDDSRVYR'))

summary(nhanes_data$LBXCOT)

```

# Save the included chems dictionary
```{r}
# add LOD % info of each chem based on included sample to dictionary
dictionary_included <- left_join(dictionary_included,
                                 above_50_pct,
                                 join_by(comment_codename_use == comment_codename))

# reorganize
dictionary_included <- dictionary_included %>%
  select(-use_chem) %>%
  relocate(n_measures, .before = n_above_lod) %>%
  relocate(unique_cycles, .after = pct_above_lod) %>%
  relocate(comment_codename_use, .after = unique_cycles)

# save dictionary
write_csv(dictionary_included, file = 'Data/use_these_chems.csv')

```

# Import cognitive data (outcome)
Using Digit Symbol Substitution Test (DSST) score
```{r}
# Load files from NHANES site
cfq1 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/1999/DataFiles/CFQ.xpt')
cfq7 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2011/DataFiles/CFQ_G.XPT')
cfq8 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2013/DataFiles/CFQ_H.XPT')

cfq_all <- bind_rows(cfq1, cfq7, cfq8)
dim(cfq_all) # 5306   26

# create overall DSST score variable from CFDRIGHT (1999-2000) and CFDDS (2011-2014)
cfq_all <- cfq_all %>%
  mutate(DSST = coalesce(CFDRIGHT, CFDDS))

summary(cfq_all$DSST) 
# range 0 - 117, 875 NAs

# pull just DSST variable to merge
dsst <- cfq_all %>%
  select(SEQN, DSST)

nhanes_data <- left_join(nhanes_data,
                         dsst,
                         by = 'SEQN')
dim(nhanes_data) # 29896    28
  
```

# Prep covariates

In main cognition ExWAS analysis:
- Age
- Sex
- Race/ethnicity
- Education
- Smoking status
- Serum cotinine
- Urinary creatinine
- Fish consumption

For sensitivity analyses:
- Estimated glomerular filtration rate
- Waist circumference
- Alcohol consumption

Other SES variables:
- Employment
- Insurance
- Housing
- Food security

## Age
```{r}
# overall
summary(nhanes_data$RIDAGEYR)

# included only
nhanes_data %>%
  filter(include == 1) %>%
  summarize(mean_age = mean(RIDAGEYR),
            median_age = median(RIDAGEYR))
```

## Sex
```{r}
# overall
nhanes_data %>%
  count(RIAGENDR)
# 1 = male, 2 = female

# included only
nhanes_data %>%
  filter(include == 1) %>%
  count(RIAGENDR)
```

## Race/ethnicity
```{r}
# overall
nhanes_data %>%
  count(RIDRETH1)
# 1 = mexican american, 2 = other hispanic, 3 = non hispanic white, 4 = non hispanic black, 5 = other race

# included only
nhanes_data %>%
  filter(include == 1) %>%
  count(RIDRETH1)
```

## Education
```{r}
# DMDEDUC2: age 20+, DMDEDUC3: age <20
# will harmonize both variables

# overall
table(nhanes_data$DMDEDUC2, useNA = 'always')
# 1 = less than 9th, 2 = some high school with no diploma, 3 = high school grad/GED, 4 = some college or AA, 5 = college grad or above, 7 = refused, 9 = don't know

table(nhanes_data$DMDEDUC3, useNA = 'always')
# 0 - 12 = grade numbers, 13 = HS graduate, 14 = GED, 15 = more than HS, 55 = less than 5th, 66 = less than 9th, 77 = refused, 99 = don't know

# convert 7/77/9/99 to NA and create binary variable: HS or above vs. below HS
nhanes_data <- nhanes_data %>%
  mutate(DMDEDUC2 = case_when(DMDEDUC2 %in% c(7, 9) ~ NA,
                              .default = DMDEDUC2),
         DMDEDUC3 = case_when(DMDEDUC3 %in% c(77, 99) ~ NA,
                              .default = DMDEDUC3),
         hs_educ = case_when(DMDEDUC2 %in% c(3:5) | DMDEDUC3 %in% c(13:15) ~ 1,
                             DMDEDUC2 %in% c(1, 2) | DMDEDUC3 %in% c(0:12, 55, 66) ~ 0,
                             .default = NA))

table(nhanes_data$hs_educ, useNA = 'always')
table(nhanes_data$hs_educ, nhanes_data$DMDEDUC2, useNA = 'always')
table(nhanes_data$hs_educ, nhanes_data$DMDEDUC3, useNA = 'always')

```

## Smoking status
```{r}
smk1 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/1999/DataFiles/SMQ.xpt') %>%
  select(SEQN, SMQ020, SMQ040)
smk7 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2011/DataFiles/SMQ_G.xpt') %>%
  select(SEQN, SMQ020, SMQ040)
smk8 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2013/DataFiles/SMQ_H.xpt') %>%
  select(SEQN, SMQ020, SMQ040)

smk_all <- bind_rows(smk1, smk7, smk8)
dim(smk_all) # 18838     3

nhanes_data <- left_join(nhanes_data, 
                         smk_all,
                         by = 'SEQN')
dim(nhanes_data) # 29896    31

table(nhanes_data$SMQ020, useNA = 'ifany')
# smoked at least 100 cigarettes in life: 1 = yes, 2 = no, 7 = refused, 9 = don't know

table(nhanes_data$SMQ040, useNA = 'ifany')
# current smoking: 1 = every day, 2 = some days, 3 = no, 7 = refused, 9 = don't know

table(nhanes_data$SMQ020, nhanes_data$SMQ040, useNA = 'ifany')

nhanes_data <- nhanes_data %>%
  mutate(smk_status = case_when(SMQ020 == 2 ~ 0, # never smokers
                                SMQ020 == 1 & SMQ040 == 3 ~ 1, # former smokers
                                SMQ020 == 1 & SMQ040 %in% c(1, 2) ~ 2, # current smokers
                                .default = NA))

table(nhanes_data$smk_status, useNA = 'always')

```

## Serum cotinine
```{r}
# overall
summary(nhanes_data$LBXCOT)

# included only
nhanes_data %>%
  filter(include == 1) %>%
  summarize(mean_cotinine = mean(LBXCOT, na.rm = T),
            median_cotinine = median(LBXCOT, na.rm = T))

```

## Urinary creatinine
```{r}
# overall
summary(nhanes_data$URXUCR)

# included only
nhanes_data %>%
  filter(include == 1) %>%
  summarize(mean_creatinine = mean(URXUCR, na.rm = T),
            median_creatinine = median(URXUCR, na.rm = T))
```

## Fish consumption
```{r}
# pull dictionary info for diet vars
diet_dictionary <- read_csv('Data/dictionary_nhanes.csv') %>%
  filter(in_dataset == 'Dietary')
# fish/shellfish consumption in the past 30 days - names start with DRD340, DRD350, DRD360, DRD370

# pull fish consumption vars
fish_dictionary <- diet_dictionary %>%
  filter(grepl('^DRD3', variable_codename_use))

fish_vars <- dietary_clean %>%
  select(SEQN, SDDSRVYR, survey_day, all_of(fish_dictionary$variable_codename_use)) %>%
  filter(SDDSRVYR %in% c(1, 7, 8))

# look at survey_day
table(fish_vars$survey_day, fish_vars$DRD340)
table(fish_vars$survey_day, fish_vars$DRD360)
# only need survey_day 1

fish_vars <- fish_vars %>%
  filter(survey_day == 1)

# vars numbered 340/350: shellfish
# 340 asks if any types shellfish eaten, 350s go thru each type if yes to 340
table(fish_vars$DRD340, useNA = 'always')

# vars numbered 360/370: fish
# 360 asks if any types fish eaten, 370s go thru each type if yes to 360
table(fish_vars$DRD360, useNA = 'always')

# remove anyone who is NA to both shellfish AND fish, then anyone who answered refused/don't know (7/9)
# makes a dataset only of participants who answered 340/360
fish_vars <- fish_vars %>%
  filter(!if_all(c(DRD340, DRD360), is.na)) %>%
  filter(!DRD340 %in% c(7, 9)) %>%
  filter(!DRD360 %in% c(7, 9))

# add up all the variables indicating # of fish/shellfish eaten
fish_scores <- fish_vars %>%
  replace(is.na(.), 0) %>% # convert any NAs to 0 - indicates none eaten
  mutate(shellfish_score = DRD350AQ + DRD350BQ + DRD350CQ + DRD350DQ + DRD350EQ + DRD350FQ + 
           DRD350GQ + DRD350HQ + DRD350IQ + DRD350JQ,
         fish_score = DRD370AQ + DRD370BQ + DRD370CQ + DRD370DQ + DRD370EQ + DRD370FQ + DRD370GQ +
           DRD370HQ + DRD370IQ + DRD370JQ + DRD370KQ + DRD370LQ + DRD370MQ + DRD370NQ + DRD370OQ + 
           DRD370PQ + DRD370QQ + DRD370RQ + DRD370SQ + DRD370TQ + DRD370UQ,
         total_seafood_score = shellfish_score + fish_score) %>%
  select(SEQN, SDDSRVYR, shellfish_score, fish_score, total_seafood_score)

summary(fish_scores$shellfish_score)
summary(fish_scores$fish_score)
summary(fish_scores$total_seafood_score)

# create categorical versions
# 0, 1-3, 4+
fish_scores <- fish_scores %>%
  mutate(shellfish_grp = case_when(shellfish_score == 0 ~ '0',
                                   shellfish_score %in% c(1:3) ~ '1-3',
                                   shellfish_score >= 4 ~ '4+'),
         fish_grp = case_when(fish_score == 0 ~ '0',
                              fish_score %in% c(1:3) ~ '1-3',
                              fish_score >= 4 ~ '4+'),
         total_seafood_grp = case_when(total_seafood_score == 0 ~ '0',
                                       total_seafood_score %in% c(1:3) ~ '1-3',
                                       total_seafood_score >= 4 ~ '4+'))

table(fish_scores$shellfish_grp, useNA = 'always')
table(fish_scores$fish_grp, useNA = 'always')
table(fish_scores$total_seafood_grp, useNA = 'always')

# add labels
fish_scores <- fish_scores %>%
  var_labels(shellfish_score = 'Number of times shellfish eaten in past 30 days',
             fish_score = 'Number of times fish eaten in past 30 days',
             total_seafood_score = 'Number of times shellfish+fish eaten in past 30 days',
             shellfish_grp = 'Categorical version of shellfish_score',
             fish_grp = 'Categorical version of fish_score',
             total_seafood_grp = 'Categorical version of total_seafood_score')

# merge scores into nhanes_data
fish_to_merge <- fish_scores %>%
  select(SEQN, SDDSRVYR, total_seafood_score, total_seafood_grp)

nhanes_data <- left_join(nhanes_data,
                         fish_to_merge,
                         by = c('SEQN', 'SDDSRVYR'))
dim(nhanes_data) # 29896    34

# check descriptives
summary(nhanes_data$total_seafood_score)
table(nhanes_data$total_seafood_grp, useNA = 'always')

# included only
nhanes_data %>%
  filter(include == 1) %>%
  summarize(mean_seafood = mean(total_seafood_score, na.rm = T),
            median_seafood = median(total_seafood_score, na.rm = T))

nhanes_data %>%
  filter(include == 1) %>%
  count(total_seafood_grp)
```

## eGFR
```{r}
# overall
summary(nhanes_data$VNEGFR)

# included only
nhanes_data %>%
  filter(include == 1) %>%
  summarize(mean_egfr = mean(VNEGFR, na.rm = T),
            median_egfr = median(VNEGFR, na.rm = T),
            nmiss = sum(is.na(VNEGFR)))

```

## Waist circumference
```{r}
# overall
summary(nhanes_data$BMXWAIST)

# included only
nhanes_data %>%
  filter(include == 1) %>%
  summarize(mean_wc = mean(BMXWAIST, na.rm = T),
            median_wc = median(BMXWAIST, na.rm = T),
            nmiss = sum(is.na(BMXWAIST))) # 370 missing

```

## Alcohol consumption
```{r}

# ALQ101: had at least 10 drinks in past year?
table(nhanes_data$ALQ101, useNA = 'always')

nhanes_data %>%
  filter(include == 1) %>%
  count(ALQ101)
# 1 yes, 2 no, 9 don't know; 270 NAs in included grp

# convert 9 to NA
nhanes_data <- nhanes_data %>%
  mutate(ALQ101 = case_when(ALQ101 == 9 ~ NA,
                            .default = ALQ101))

table(nhanes_data$ALQ101, useNA = 'always')

# ALQ120Q: alcohol consumption frequency in past 12 months; ALQ120U: units to match ALQ120Q 
table(nhanes_data$ALQ120Q, useNA = 'always')
# convert ALQ120Q = 777 and 999 to NA
# create new number of drinks/12 months var: if ALQ101 = 2 (no), then set to 0 drinks - otherwise ALQ120Q
nhanes_data <- nhanes_data %>%
  mutate(ALQ120Q = case_when(ALQ120Q %in% c(777, 999) ~ NA,
                             .default = ALQ120Q),
         drinks_past_yr = case_when(ALQ101 == 2 ~ 0,
                                    .default = ALQ120Q))

# based on units from ALQ120U, create multiplier to convert to drinks per month
table(nhanes_data$ALQ120U, useNA = 'always')
# 1 = week, 2 = month, 3 = year
# if week: multiply by 52weeks/12months = 4.33
# if month: multiply by 1
# if year: multiply by 1year/12months = 0.083

nhanes_data <- nhanes_data %>%
  mutate(alc_unit_mult = case_when(ALQ120U == 1 ~ 4.33,
                                   ALQ120U == 2 ~ 1,
                                   ALQ120U == 3 ~ 0.083,
                                   drinks_past_yr == 0 ~ 1)) # fill in a 1 if no alc consumed, otherwise it would try to mult by NA

table(nhanes_data$alc_unit_mult, useNA = 'always')

# calculate number of drinks per month
nhanes_data <- nhanes_data %>%
  mutate(alc_month_freq = drinks_past_yr * alc_unit_mult)

summary(nhanes_data$alc_month_freq, useNA = 'always')

# confirm that no NAs in participants with 0 drinks
alq_0_drinks <- nhanes_data %>%
  filter(drinks_past_yr == 0)

summary(alq_0_drinks$alc_month_freq) # all 0, good

# create binary alcohol intake (0 if <=4/month, 1 if > 4/month)
nhanes_data <- nhanes_data %>%
  mutate(alc4 = case_when(alc_month_freq <= 4 ~ 0,
                          alc_month_freq > 4 ~ 1))

table(nhanes_data$alc4, useNA = 'always')

# check included group
nhanes_data %>%
  filter(include == 1) %>%
  count(alc4) # 280 NAs

```

## Health insurance
```{r}
# Note that questions/variables are different between 1999-2000 and 2011-2014
hiq1 = read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/1999/DataFiles/HIQ.xpt')
hiq7 = read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2011/DataFiles/HIQ_G.xpt')
hiq8 = read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2013/DataFiles/HIQ_H.xpt')

# clean 1999-2000 first
hiq1 <- hiq1 %>%
  select(SEQN, HID010, HID030A, HID030B, HID030C, HID030D)

table(hiq1$HID010, useNA = 'always')
# 1 = yes ins, 2 = no ins, 7/9/NA
# 2038 uninsured

table(hiq1$HID030A, useNA = 'always') # 5215 yes private ins
table(hiq1$HID030B, useNA = 'always') # 1232 yes medicare
table(hiq1$HID030C, useNA = 'always') # 1492 yes medicaid or chip
table(hiq1$HID030D, useNA = 'always') # 534 yes other govt ins

# yes (1) to HID030A = private
# yes (1) to B, C, or D = govt
# participants can answer Y to more than 1 type - deciding to prioritize Y to private over govt types for categorization

hiq1 <- hiq1 %>%
  mutate(ins_status = case_when(HID010 == 2 ~ 1, # no ins
                                HID030A == 1 ~ 0, # private ins
                                (HID030B == 1 | HID030C == 1 | HID030D == 1) ~ 0.5)) # govt ins

table(hiq1$ins_status, useNA = 'always')
# 266 NAs

# 2011-2014 next
hiq_7_8 = rbind(hiq7, hiq8) %>%
    select(SEQN, HIQ011, HIQ031A, HIQ031B, HIQ031C, HIQ031D, HIQ031E, HIQ031F, HIQ031G, HIQ031H, HIQ031I) 

table(hiq_7_8$HIQ011, useNA = 'always') # no insurance (2) = 3273
table(hiq_7_8$HIQ031A, useNA = 'always') # private ins (14) = 9154
table(hiq_7_8$HIQ031B, useNA = 'always') # medicare (15) = 2432
table(hiq_7_8$HIQ031C, useNA = 'always') # medi-gap (16) = 25
table(hiq_7_8$HIQ031D, useNA = 'always') # medicaid (17) = 4577
table(hiq_7_8$HIQ031E, useNA = 'always') # chip (18) = 151
table(hiq_7_8$HIQ031F, useNA = 'always') # military (19) = 402
table(hiq_7_8$HIQ031G, useNA = 'always') # indian health service (20) = 21
table(hiq_7_8$HIQ031H, useNA = 'always') # state sponsored (21) = 1155
table(hiq_7_8$HIQ031I, useNA = 'always') # other govt (22) = 505

# from documentation of 05-06 questionnaire (old = new):
# HID010 = HIQ011
# private ins HID030A = HIQ031A (private) and HIQ031C (medigap)
# govt ins HID030B, HID030C, HID030D = HIQ031B, HIQ031D thru I (covers medicare, medicaid, chip, military, IHS, state sponsored, and other govt ins)

hiq_7_8 <- hiq_7_8 %>%
  mutate(ins_status = case_when(HIQ011 == 2 ~ 1, # no ins
                                (HIQ031A == 14 | HIQ031C == 16) ~ 0, # private ins
                                (HIQ031B == 15 | HIQ031D == 17 | HIQ031E == 18 | HIQ031F == 19 | 
                                   HIQ031G == 20 | HIQ031H == 21 | HIQ031I == 22) ~ 0.5)) # govt ins

table(hiq_7_8$ins_status, useNA = 'always') # 116 NAs

# combine the two sets
hiq1_keep <- hiq1 %>%
  select(SEQN, ins_status)
hiq_7_8_keep <- hiq_7_8 %>%
  select(SEQN, ins_status)

hiq_all <- bind_rows(hiq1_keep, hiq_7_8_keep)

table(hiq_all$ins_status, useNA = 'always')

nhanes_data <- left_join(nhanes_data,
                         hiq_all,
                         by = 'SEQN')

table(nhanes_data$ins_status, useNA = 'always')

```

## Housing tenure
```{r}
hoq1 = read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/1999/DataFiles/HOQ.xpt')
hoq7 = read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2011/DataFiles/HOQ_G.xpt')
hoq8 = read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2013/DataFiles/HOQ_H.xpt')

# HOQ065 = housing tenure, own/rent/other
hoq_all = bind_rows(hoq1, hoq7, hoq8) %>%
  select(SEQN, HOQ065)

nhanes_data = left_join(nhanes_data,
                        hoq_all,
                        by = 'SEQN')

table(nhanes_data$HOQ065, useNA = 'always')

# convert 7 and 9 to NA
nhanes_data <- nhanes_data %>%
  mutate(HOQ065 = case_when(HOQ065 == 7 ~ NA,
                            HOQ065 == 9 ~ NA,
                            .default = HOQ065))

table(nhanes_data$HOQ065, useNA = 'always')

```

## Food security
```{r}
fsq1 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/1999/DataFiles/FSQ.xpt')
fsq7 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2011/DataFiles/FSQ_G.xpt')
fsq8 <- read_xpt('https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2013/DataFiles/FSQ_H.xpt')

fsq_all <- bind_rows(fsq1, fsq7, fsq8) %>%
  select(SEQN, ADFDSEC, FSDAD)
# 1999-2000 uses ADFDSEC, 2011-2014 uses FSDAD

# merge into sample dataset
nhanes_data <- left_join(nhanes_data,
                         fsq_all,
                         by = 'SEQN')

# harmonize variables across cycles
nhanes_data <- nhanes_data %>%
  mutate(ad_foodsec = coalesce(ADFDSEC, FSDAD))
# 1 = full security, 2 = marginal, 3 = low, 4 = very low

table(nhanes_data$ad_foodsec, useNA = 'always')

```

## Employment
```{r}
table(nhanes_data$OCD150, useNA = 'always')
table(nhanes_data$OCQ380, useNA = 'always')

# create binary variable for employment status
nhanes_data <- nhanes_data %>%
  mutate(employed = case_when(OCD150 %in% c(1, 2) ~ 1, # employed
                              OCD150 == 3 | OCQ380 %in% c(1:7) ~ 0, # not employed (incl. retired)
                              .default = NA))

table(nhanes_data$employed, useNA = 'always')
table((nhanes_data %>% filter(include == 1))$employed, useNA = 'always')

```

# Remove unnecessary variables from nhanes_data
```{r}
dim(nhanes_data) # 29896    44
colnames(nhanes_data)

cognition_ewas_to_impute <- nhanes_data %>%
  select(-c('SEQN_new', 'OCD150', 'OCQ380', 'ALQ101', 'ALQ120Q', 'ALQ120U', 'aged_60plus',
            'SMQ020', 'SMQ040', 'drinks_past_yr', 'alc_unit_mult', 'ADFDSEC', 'FSDAD'))
dim(cognition_ewas_to_impute) # 29896    31

```

# Add variable labels
```{r}

get_label(cognition_ewas_to_impute)

# fill in blank variable labels
cognition_ewas_to_impute <- cognition_ewas_to_impute %>%
  var_labels(include = 'Included in analysis',
             DSST = 'DSST score (all cycles)',
             hs_educ = 'Education (high school completion)',
             smk_status = 'Smoking status',
             alc_month_freq = 'Alcohol consumption (drinks per month)',
             alc4 = 'Alcohol consumption (Over 4 drinks per month)',
             ins_status = 'Health insurance coverage type',
             HOQ065 = 'Home owned, bought, rented, or other',
             ad_foodsec = 'Food security level (all cycles)',
             employed = 'Employment status')

# add value labels for categorical variables
cognition_ewas_to_impute <- cognition_ewas_to_impute %>%
  val_labels(SDDSRVYR = c('1999-2000', 
                          '2011-2012', 
                          '2013-2014'),
             RIAGENDR = c('Male',
                          'Female'),
             RIDRETH1 = c('Mexican American',
                          'Other Hispanic',
                          'Non-Hispanic White',
                          'Non-Hispanic Black',
                          'Other Race'),
             include = c('Excluded',
                         'Included'),
             hs_educ = c('Did not complete high school',
                         'Completed high school or above'),
             smk_status = c('Never smoker',
                            'Former smoker',
                            'Current smoker'),
             alc4 = c('0-4 drinks per month',
                      '>4 drinks per month'),
             ins_status = c('Private insurance',
                            'Government insurance',
                            'No insurance'),
             ad_foodsec = c('Full security',
                            'Marginal security',
                            'Low security',
                            'Very low security'),
             employed = c('Unemployed or retired',
                          'Employed'))

str(cognition_ewas_to_impute)

```

# Save dataset to be imputed
```{r}
saveRDS(cognition_ewas_to_impute, file = 'Data/cognition_ewas_to_impute.RDS')

```
