==> 5_Cognition_ExWAS_volcano_plots_github_Feb_14_2025.Rmd <==
---
title: "Cognition ExWAS plots"
date: "`r Sys.Date()`"
output: html_document
---

Purpose: create volcano plots for cognition ExWAS analyses

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sjlabelled)
library(ggrepel)
library(ggpubr)
```

# Load results files
```{r}
# main analysis
main <- read_csv('Output Cog ExWAS/Cog ExWAS main analysis model output chem term Jul 9 2024.csv')

# normal eGFR
normal_egfr <- read_csv('Output Cog ExWAS/Cog ExWAS normal eGFR model output chem term Jul 9 2024.csv')

# add lifestyle covariates
add_covars <- read_csv('Output Cog ExWAS/Cog ExWAS lifestyle model output chem term Jul 9 2024.csv')

# binary mild cognitive impairment
poisson <- read_csv('Output Cog ExWAS/Cog ExWAS poisson model output chem term Nov 4 2024.csv')

# binary, normal eGFR
poisson_egfr <- read_csv('Output Cog ExWAS/Cog ExWAS poisson normal eGFR model output chem term Nov 4 2024.csv')

# sex stratified
sex_stratified <- read_csv('Output Cog ExWAS/Cog ExWAS sex stratified output chem term Oct 25 2024.csv')

```

# Set up chemical family colors
```{r}
# sort chemical families in alpha order
chem_family_alpha <- main %>%
  select(chemical_family) %>%
  arrange(chemical_family) %>%
  unique()

chem_family_order <- chem_family_alpha$chemical_family

# define colors
chem_fam_colors <- c("#8B0000", # Acrylamide
                     "#0E1171", # Aldehydes
                     "#be67c9", # Aromatic amines
                     "#FFA500", # Dioxins
                     "#228B22", # Metals
                     "#FFB6C1", # PFAS
                     "#A2CD5A", # Personal Care
                     "#1E90FF", # Pesticides
                     "#FF6B00", # PFRs
                     "#A4D3EE", # Phthalates
                     "#cf9b76", # PAHs
                     "#FF69B4", # PCBs
                     "#8B4513", # Smoking related compounds
                     "#828282") # VOCs

```

# Main analysis
```{r}
# factor chem family
main <- main %>%
  mutate(chemical_family = factor(chemical_family,
                                  levels = chem_family_order)) 

# add labels
main <- main %>%
  mutate(chem_label = case_when(p_val_unadj < 0.05 ~ gsub('\\s\\(([^()]+)\\)$', # remove units ' (__)' from the end of the name
                                                          '',
                                                          full_chem_name),
                                .default = ''))

# plot
volcano_plot_cog_exwas <- ggplot(main, 
                                 aes(x = estimate,
                                     y = -log10(p_val_unadj), 
                                     label = chem_label,
                                     color = chemical_family)) +
  theme_light() +
  # p-value lines and labels
  geom_hline(yintercept = -log10(0.05),
             linetype = 'dashed',
             color = 'red') + 
  geom_hline(yintercept = -log10(0.01),
             linetype = 'dashed',
             color = 'blue') +
  annotate('text', x = -4.1, y = -log10(0.035), size = 12/.pt, label = 'p = 0.05') +
  annotate('text', x = -4.1, y = -log10(0.007), size = 12/.pt, label = 'p = 0.01') +
  # vertical line
  geom_vline(xintercept = 0,
             linetype = 'dashed',
             color = 'black') +
  # top arrows and labels
  annotate('segment', x = -0.1, y = 6, xend = -0.6, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = -1, y = 6, size = 12/.pt, label = 'Lower cognition') +
  annotate('segment', x = 0.1, y = 6, xend = 0.6, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 1, y = 6, size = 12/.pt, label = 'Higher cognition') +
  geom_point(size = 3) +
  geom_text_repel(data = subset(main, p_val_unadj < 0.01),
                  size = 10/.pt,
                  nudge_x = 0.1) +
  geom_text_repel(data = subset(main, p_val_unadj > 0.01),
                  size = 10/.pt,
                  nudge_y = -0.1) +
  scale_color_manual(values = chem_fam_colors) +
  xlim(-4.1, 2.5) +
  ylim(-0.1, 6) +
  xlab(expression('Difference in cognition score per 1 standard deviation increase of the log'[2] ~ 'scaled chemical measure')) +
  ylab(expression('-log'[10] ~ '(p value)')) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 12))

volcano_plot_cog_exwas 

```

## Save 
```{r}

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS main analysis volcano plot Oct 16 2024.pdf',
       plot = volcano_plot_cog_exwas,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS main analysis volcano plot Oct 16 2024.png',
       plot = volcano_plot_cog_exwas,
       height = 9,
       width = 14)

```


# Normal eGFR subset
```{r}
# factor chem family
normal_egfr <- normal_egfr %>%
  mutate(chemical_family = factor(chemical_family,
                                  levels = chem_family_order)) 

# add labels
normal_egfr <- normal_egfr %>%
  mutate(chem_label = case_when(p_val_unadj < 0.05 ~ gsub('\\s\\(([^()]+)\\)$',
                                                          '',
                                                          full_chem_name),
                                .default = ''))

# plot
volcano_plot_cog_exwas_egfr <- ggplot(normal_egfr, 
                                      aes(x = estimate,
                                          y = -log10(p_val_unadj), 
                                          label = chem_label,
                                          color = chemical_family)) +
  theme_light() +
  geom_hline(yintercept = -log10(0.05),
             linetype = 'dashed',
             color = 'red') + 
  geom_hline(yintercept = -log10(0.01),
             linetype = 'dashed',
             color = 'blue') +
  annotate('text', x = -4.1, y = -log10(0.035), size = 12/.pt, label = 'p = 0.05') +
  annotate('text', x = -4.1, y = -log10(0.007), size = 12/.pt, label = 'p = 0.01') +
  geom_vline(xintercept = 0,
             linetype = 'dashed',
             color = 'black') +
  annotate('segment', x = -0.1, y = 5, xend = -0.6, yend = 5, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = -1, y = 5, size = 12/.pt, label = 'Lower cognition') +
  annotate('segment', x = 0.1, y = 5, xend = 0.6, yend = 5, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 1, y = 5, size = 12/.pt, label = 'Higher cognition') +
  geom_point(size = 3) +
  geom_text_repel(data = subset(normal_egfr,
                                p_val_unadj < 0.05),
                  size = 10/.pt,
                  nudge_x = 0.1) + 
  scale_color_manual(values = chem_fam_colors) +
  xlim(-4.1, 2.5) +
  ylim(-0.1, 5) +
  xlab(expression('Difference in cognition score per 1 standard deviation increase of the log'[2] ~ 'scaled chemical measure, subsetted to participants with normal eGFR')) +
  ylab(expression('-log'[10] ~ '(p value)')) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 12))

volcano_plot_cog_exwas_egfr 

```

## Save
```{r}

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS normal eGFR volcano plot Oct 16 2024.pdf',
       plot = volcano_plot_cog_exwas_egfr,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS normal eGFR volcano plot Oct 16 2024.png',
       plot = volcano_plot_cog_exwas_egfr,
       height = 9,
       width = 14)

```

# Adding lifestyle covariates
```{r}
# factor chem family
add_covars <- add_covars %>%
  mutate(chemical_family = factor(chemical_family,
                                  levels = chem_family_order)) 

# add labels
add_covars <- add_covars %>%
  mutate(chem_label = case_when(p_val_unadj < 0.05 ~ gsub('\\s\\(([^()]+)\\)$',
                                                          '',
                                                          full_chem_name),
                                .default = ''))

# plot
volcano_plot_cog_exwas_lifestyle <- ggplot(add_covars, 
                                      aes(x = estimate,
                                          y = -log10(p_val_unadj), 
                                          label = chem_label,
                                          color = chemical_family)) +
  theme_light() +
  geom_hline(yintercept = -log10(0.05),
             linetype = 'dashed',
             color = 'red') + 
  geom_hline(yintercept = -log10(0.01),
             linetype = 'dashed',
             color = 'blue') +
  # annotate lines
  annotate('text', x = -4.1, y = -log10(0.035), size = 12/.pt, label = 'p = 0.05') +
  annotate('text', x = -4.1, y = -log10(0.007), size = 12/.pt, label = 'p = 0.01') +
  geom_vline(xintercept = 0,
             linetype = 'dashed',
             color = 'black') +
  annotate('segment', x = -0.1, y = 5, xend = -0.6, yend = 5, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = -1, y = 5, size = 12/.pt, label = 'Lower cognition') +
  annotate('segment', x = 0.1, y = 5, xend = 0.6, yend = 5, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 1, y = 5, size = 12/.pt, label = 'Higher cognition') +
  geom_point(size = 3) +
  geom_text_repel(size = 10/.pt,
                  max.overlaps = 20) + 
  scale_color_manual(values = chem_fam_colors) +
  xlim(-4.1, 2.5) +
  ylim(-0.1, 5) +
  xlab(expression('Difference in cognition score per 1 standard deviation increase of the log'[2] ~ 'scaled chemical measure, adjusted further for waist circumference and alcohol consumption')) +
  ylab(expression('-log'[10] ~ '(p value)')) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 12))

volcano_plot_cog_exwas_lifestyle

```

## Save
```{r}

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS lifestyle covars volcano plot Oct 16 2024.pdf',
       plot = volcano_plot_cog_exwas_lifestyle,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS lifestyle covars volcano plot Oct 16 2024.png',
       plot = volcano_plot_cog_exwas_lifestyle,
       height = 9,
       width = 14)

```

# Binary MCI
```{r}
# factor chem family
poisson <- poisson %>%
  mutate(chemical_family = factor(chemical_family,
                                  levels = chem_family_order)) 

# add labels
poisson <- poisson %>%
  mutate(chem_label = case_when(p_val_unadj < 0.05 ~ gsub('\\s\\(([^()]+)\\)$',
                                                          '',
                                                          full_chem_name),
                                .default = ''))

# plot
volcano_plot_cog_exwas_poisson <- ggplot(poisson, 
                                         aes(x = exp_est,
                                             y = -log10(p_val_unadj), 
                                             label = chem_label,
                                             color = chemical_family)) +
  theme_light() +
  geom_hline(yintercept = -log10(0.05),
             linetype = 'dashed',
             color = 'red') + 
  geom_hline(yintercept = -log10(0.01),
             linetype = 'dashed',
             color = 'blue') +
  # annotate lines
  annotate('text', x = 0.6, y = -log10(0.035), size = 12/.pt, label = 'p = 0.05') +
  annotate('text', x = 0.6, y = -log10(0.007), size = 12/.pt, label = 'p = 0.01') +
  geom_vline(xintercept = 1,
             linetype = 'dashed',
             color = 'black') +
  annotate('segment', x = 0.95, y = 6, xend = 0.85, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 0.75, y = 6, size = 12/.pt, label = 'Lower risk of cognitive impairment') +
  annotate('segment', x = 1.05, y = 6, xend = 1.15, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 1.3, y = 6, size = 12/.pt, label = 'Higher risk of cognitive impairment') +
  geom_point(size = 3) +
  geom_text_repel(size = 10/.pt,
                  max.overlaps = 20,
                  nudge_y = 0.1) + 
  scale_color_manual(values = chem_fam_colors) +
  scale_x_continuous(trans = 'log10',
                     limits = c(0.6, 1.6),
                     breaks = seq(0.6, 1.6, by = 0.2)) +
  ylim(-0.1, 6) +
  xlab(expression('Relative risk of mild cognitive impairment per 1 standard deviation increase of the log'[2] ~ 'scaled chemical measure')) +
  ylab(expression('-log'[10] ~ '(p value)')) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 12))

volcano_plot_cog_exwas_poisson

```

## Save
```{r}

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS poisson volcano plot Nov 4 2024.pdf',
       plot = volcano_plot_cog_exwas_poisson,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS poisson volcano plot Nov 4 2024.png',
       plot = volcano_plot_cog_exwas_poisson,
       height = 9,
       width = 14)

```

# Binary MCI, normal eGFR
```{r}
# factor chem family
poisson_egfr <- poisson_egfr %>%
  mutate(chemical_family = factor(chemical_family,
                                  levels = chem_family_order)) 

# add labels
poisson_egfr <- poisson_egfr %>%
  mutate(chem_label = case_when(p_val_unadj < 0.05 ~ gsub('\\s\\(([^()]+)\\)$', # remove units ' (__)' from the end of the name
                                                          '',
                                                          full_chem_name),
                                .default = ''))

# plot
volcano_plot_cog_exwas_poisson_egfr <- ggplot(poisson_egfr, 
                                         aes(x = exp_est,
                                             y = -log10(p_val_unadj), 
                                             label = chem_label,
                                             color = chemical_family)) +
  theme_light() +
  geom_hline(yintercept = -log10(0.05),
             linetype = 'dashed',
             color = 'red') + 
  geom_hline(yintercept = -log10(0.01),
             linetype = 'dashed',
             color = 'blue') +
  # annotate lines
  annotate('text', x = 0.5, y = -log10(0.035), size = 12/.pt, label = 'p = 0.05') +
  annotate('text', x = 0.5, y = -log10(0.007), size = 12/.pt, label = 'p = 0.01') +
  geom_vline(xintercept = 1,
             linetype = 'dashed',
             color = 'black') +
  annotate('segment', x = 0.95, y = 4, xend = 0.85, yend = 4, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 0.73, y = 4, size = 12/.pt, label = 'Lower risk of cognitive impairment') +
  annotate('segment', x = 1.05, y = 4, xend = 1.17, yend = 4, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 1.35, y = 4, size = 12/.pt, label = 'Higher risk of cognitive impairment') +
  geom_point(size = 3) +
  geom_text_repel(size = 10/.pt,
                  max.overlaps = 20,
                  nudge_y = 0.1) + 
  scale_color_manual(values = chem_fam_colors) +
  scale_x_continuous(trans = 'log10',
                     limits = c(0.5, 1.8),
                     breaks = seq(0.6, 1.8, by = 0.2)) +
  ylim(-0.1, 4) +
  xlab(expression('Relative risk of mild cognitive impairment per 1 standard deviation increase of the log'[2] ~ 'scaled chemical measure, subsetted to participants with normal eGFR')) +
  ylab(expression('-log'[10] ~ '(p value)')) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 12))

volcano_plot_cog_exwas_poisson_egfr

```

## Save
```{r}

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS normal eGFR poisson volcano plot Nov 4 2024.pdf',
       plot = volcano_plot_cog_exwas_poisson_egfr,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS normal eGFR poisson volcano plot Nov 4 2024.png',
       plot = volcano_plot_cog_exwas_poisson_egfr,
       height = 9,
       width = 14)

```

# Sex stratified
## Males
```{r}
# separate results
males <- sex_stratified %>%
  filter(sex == 'male')

# factor chem family
males <- males %>%
  mutate(chemical_family = factor(chemical_family,
                                  levels = chem_family_order)) 

# add labels
males <- males %>%
  mutate(chem_label = case_when(p_val_unadj < 0.05 ~ gsub('\\s\\(([^()]+)\\)$', # remove units ' (__)' from the end of the name
                                                          '',
                                                          full_chem_name),
                                .default = ''))

# plot
volcano_plot_males <- ggplot(males, 
                             aes(x = estimate,
                                 y = -log10(p_val_unadj), 
                                 label = chem_label,
                                 color = chemical_family)) +
  theme_light() +
  # p-value lines and labels
  geom_hline(yintercept = -log10(0.05),
             linetype = 'dashed',
             color = 'red') + 
  geom_hline(yintercept = -log10(0.01),
             linetype = 'dashed',
             color = 'blue') +
  annotate('text', x = -5, y = -log10(0.035), size = 12/.pt, label = 'p = 0.05') +
  annotate('text', x = -5, y = -log10(0.007), size = 12/.pt, label = 'p = 0.01') +
  # vertical line
  geom_vline(xintercept = 0,
             linetype = 'dashed',
             color = 'black') +
  # top arrows and labels
  annotate('segment', x = -0.1, y = 6, xend = -0.6, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = -1.5, y = 6, size = 12/.pt, label = 'Lower cognition') +
  annotate('segment', x = 0.1, y = 6, xend = 0.6, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 1.5, y = 6, size = 12/.pt, label = 'Higher cognition') +
  geom_point(size = 3) +
  geom_text_repel(data = subset(males, p_val_unadj < 0.01),
                  size = 10/.pt,
                  nudge_x = 0.1) +
  geom_text_repel(data = subset(males, p_val_unadj > 0.01),
                  size = 10/.pt,
                  nudge_y = -0.1) +
  scale_color_manual(values = chem_fam_colors) +
  xlim(-5, 5) +
  ylim(-0.1, 6) +
  xlab(expression(paste(Difference~'in'~cognition~score~per~1~standard~deviation~increase~of~the~log[2]~scaled~chemical~measure, ', ', males))) +
  ylab(expression('-log'[10] ~ '(p value)')) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 10))

volcano_plot_males 


```

## Females
```{r}
# separate results
females <- sex_stratified %>%
  filter(sex == 'female')

# factor chem family
females <- females %>%
  mutate(chemical_family = factor(chemical_family,
                                  levels = chem_family_order)) 

# add labels
females <- females %>%
  mutate(chem_label = case_when(p_val_unadj < 0.05 ~ gsub('\\s\\(([^()]+)\\)$', # remove units ' (__)' from the end of the name
                                                          '',
                                                          full_chem_name),
                                .default = ''))

# plot
volcano_plot_females <- ggplot(females, 
                             aes(x = estimate,
                                 y = -log10(p_val_unadj), 
                                 label = chem_label,
                                 color = chemical_family)) +
  theme_light() +
  # p-value lines and labels
  geom_hline(yintercept = -log10(0.05),
             linetype = 'dashed',
             color = 'red') + 
  geom_hline(yintercept = -log10(0.01),
             linetype = 'dashed',
             color = 'blue') +
  annotate('text', x = -5, y = -log10(0.035), size = 12/.pt, label = 'p = 0.05') +
  annotate('text', x = -5, y = -log10(0.007), size = 12/.pt, label = 'p = 0.01') +
  # vertical line
  geom_vline(xintercept = 0,
             linetype = 'dashed',
             color = 'black') +
  # top arrows and labels
  annotate('segment', x = -0.1, y = 6, xend = -0.6, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = -1.5, y = 6, size = 12/.pt, label = 'Lower cognition') +
  annotate('segment', x = 0.1, y = 6, xend = 0.6, yend = 6, arrow = arrow(type = 'closed', length = unit(0.1, 'inches'))) +
  annotate('text', x = 1.5, y = 6, size = 12/.pt, label = 'Higher cognition') +
  geom_point(size = 3) +
  geom_text_repel(data = subset(females, p_val_unadj < 0.01),
                  size = 10/.pt,
                  nudge_x = 0.1) +
  geom_text_repel(data = subset(females, p_val_unadj > 0.01),
                  size = 10/.pt,
                  nudge_y = -0.1) +
  scale_color_manual(values = chem_fam_colors) +
  xlim(-5, 5) +
  ylim(-0.1, 6) +
  xlab(expression(paste(Difference~'in'~cognition~score~per~1~standard~deviation~increase~of~the~log[2]~scaled~chemical~measure, ', ', females))) +
  ylab(expression('-log'[10] ~ '(p value)')) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 10))

volcano_plot_females 

```

## Combine panels
```{r}
volcano_plots_by_sex <- ggarrange(volcano_plot_males, volcano_plot_females,
                                  nrow = 1, ncol = 2,
                                  common.legend = T, legend = 'bottom')

volcano_plots_by_sex

```

## Save
```{r}
# Males
ggsave(filename = 'Output Cog ExWAS/Cog ExWAS males stratified volcano plot Oct 24 2024.pdf',
       plot = volcano_plot_males,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS males stratified volcano plot Oct 24 2024.png',
       plot = volcano_plot_males,
       height = 9,
       width = 14)

# Females
ggsave(filename = 'Output Cog ExWAS/Cog ExWAS females stratified volcano plot Oct 24 2024.pdf',
       plot = volcano_plot_females,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS females stratified volcano plot Oct 24 2024.png',
       plot = volcano_plot_females,
       height = 9,
       width = 14)

# Combined
ggsave(filename = 'Output Cog ExWAS/Cog ExWAS sex stratified volcano plot combo Oct 25 2024.pdf',
       plot = volcano_plots_by_sex,
       height = 9,
       width = 16)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS sex stratified volcano plot combo Oct 25 2024.png',
       plot = volcano_plots_by_sex,
       height = 9,
       width = 16)

```


# Additional supplemental plot: forest plot of all main analysis results
```{r}
# remove units from names
main$full_chem_name <- gsub('\\s\\(([^()]+)\\)$',
                            '',
                            main$full_chem_name)

forest_plot <- ggplot(data = main,
                      aes(x = estimate,
                          y = reorder(full_chem_name, -estimate), # sort by direction of estimate
                          xmin = ci_lower,
                          xmax = ci_upper,
                          color = chemical_family)) +
  theme_light() +
  geom_vline(xintercept = 0,
             linetype = 2) +
  geom_pointrange(aes(col = chemical_family),
                  position = position_dodge2(width = 1),
                  size = 0.3) +
  scale_color_manual(values = chem_fam_colors) +
  theme(axis.text.x = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.text.y = element_text(size = 3.5)) +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 9)) +
  xlab(expression('Difference in cognition score per 1 standard deviation increase of the log'[2] ~ 'scaled chemical measure')) +
  ylab('') +
  scale_x_continuous(limits = c(-6, 6),
                     breaks = c(-6, -4, -2, 0, 2, 4, 6))

forest_plot

```

## Save
```{r}

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS main analysis forest plot Nov 21 2024.pdf',
       plot = forest_plot,
       height = 9,
       width = 14)

ggsave(filename = 'Output Cog ExWAS/Cog ExWAS main analysis forest plot Nov 21 2024.png',
       plot = forest_plot,
       height = 9,
       width = 14)

```
