==> 6_Chemical_exposure_figures_github_Feb_14_2025.Rmd <==
---
title: "Chemical stat figures"
date: "`r Sys.Date()`"
output: html_document
---

Purpose: create figures for chemical exposures (scatterplot of % above LOD stats for each chemical, histograms of chemicals, correlation matrix)

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggrepel)
library(survey)
library(dichromat)
library(pheatmap)
```

# Scatterplot- N measures vs % above LOD
## Load descriptive stats output
```{r}
chem_stats <- read_csv('Output Cog ExWAS/Chem descriptive stats weighted Jul 9 2024.csv')

# remove units from chem names
chem_stats$variable_description_use <- gsub('\\s\\(([^()]+)\\)$', 
                                            '',
                                            chem_stats$variable_description_use)
```

## Plot
```{r}
# sort chemical families in alpha order
chem_family_alpha <- chem_stats %>%
  select(chemical_family) %>%
  arrange(chemical_family) %>%
  unique()

chem_family_order <- chem_family_alpha$chemical_family

# apply order to chem_stats
chem_stats$chemical_family <- factor(chem_stats$chemical_family,
                                     levels = chem_family_order)

# assign colors to chem families
chem_fam_colors <- c("#8B0000", # Acrylamide
                     "#0E1171", # Aldehydes
                     "#be67c9", # Aromatic amines
                     "#FFA500", # Dioxins
                     "#228B22", # Metals
                     "#FFB6C1", # PFAS
                     "#A2CD5A", # Personal Care
                     "#1E90FF", # Pesticides
                     "#FF6B00", # PFRs
                     "#A4D3EE", # Phthalates
                     "#cf9b76", # PAHs
                     "#FF69B4", # PCBs
                     "#8B4513", # Smoking related compounds
                     "#828282") # VOCs

# plot
scatter_partic_by_chem_LOD <- ggplot(chem_stats,
                                     aes(x = unwt_n,
                                         y = pct_above_lod,
                                         color = chemical_family)) +
  geom_point(size = 3) +
  scale_color_manual(name = 'Chemical Family', 
                     values = chem_fam_colors) +
  geom_text_repel(aes(label = ifelse(unwt_n > 2000, 
                                     as.character(variable_description_use), '')),
                  size = 10/.pt) +
  geom_text_repel(aes(label = ifelse(pct_above_lod < 55 & unwt_n < 10000,
                                     as.character(variable_description_use), '')),
                  size = 10/.pt) +
  labs(x = 'Participants',
       y = 'Percent Above LOD (%)') +
  theme_light() +
  theme(legend.position = 'bottom',
        legend.title = element_blank(),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 12))

scatter_partic_by_chem_LOD

```

## Save
```{r}
# scatterplot
ggsave(filename = 'Output Cog ExWAS/Scatterplot of chem N and LOD Oct 22 2024.pdf',
       plot = scatter_partic_by_chem_LOD,
       width = 14,
       height = 9)

ggsave(filename = 'Output Cog ExWAS/Scatterplot of chem N and LOD Oct 22 2024.png',
       plot = scatter_partic_by_chem_LOD,
       width = 14,
       height = 9,
       dpi = 600)

```

# Histograms
## Load chem data
```{r}
# pull chems significant in main analyses
chems <- read_csv('Data/chemicals_clean.csv', na = '.') %>%
  filter(SDDSRVYR %in% c(1, 7, 8))

sig_chems <- c('LBXCOT', 'URXUTU', 'LBXVXY', 'URXBP3', 'URX2MH',
               'URXMZP', 'URX34M', 'LBXPFOA')

incl_chems <- chems %>%
  select(SEQN, SDDSRVYR, all_of(sig_chems))

# pull incl/excl and survey vars from main data
ewas_data <- readRDS('Data/cognition_ewas_to_impute.RDS') %>%
  select(SEQN, SDDSRVYR,SDMVPSU, SDMVSTRA, WTMEC2YR, include) %>%
  mutate(wt_adj = (1/3)*WTMEC2YR)

# merge
ewas_data_chems <- left_join(ewas_data,
                             incl_chems,
                             join_by(SEQN, SDDSRVYR))

```

## Set up survey object
```{r}
ewas_wt <- svydesign(ids = ~SDMVPSU, 
                     strata = ~SDMVSTRA,
                     weights = ~wt_adj,
                     nest = T,
                     data = ewas_data_chems)

ewas_wt_subset <- subset(ewas_wt,
                         include == 1)

```

## Plot in a grid
```{r}
# stop scientific notation for axes
options(scipen = 999)

# build plot in base graphics
pdf(file = 'Output Cog ExWAS/Cog ExWAS chem histograms weighted Dec 4 2024.pdf',
    height = 9, width = 14)

par(mfrow = c(3, 3),
    oma = c(2, 4, 1, 1))

LBXCOT_hist <- svyhist(~scale(log2(LBXCOT)), ewas_wt_subset, probability = F, breaks = 30,
                       las = 1, xlab = '', ylab = '', main = 'Serum cotinine')

LBXPFOA_hist <- svyhist(~scale(log2(LBXPFOA)), ewas_wt_subset, probability = F, breaks = 40,
                        las = 1, xlab = '', ylab = '', main = 'Perfluorooctanoic acid')

LBXVXY_hist <- svyhist(~scale(log2(LBXVXY)), ewas_wt_subset, probability = F, breaks = 30,
                       las = 1, xlab = '', ylab = '', main = 'Blood m/p-xylene')

URX2MH_hist <- svyhist(~scale(log2(URX2MH)), ewas_wt_subset, probability = F, breaks = 30,
                       las = 1, xlab = '', ylab = '', main = '2-Methylhippuric acid')

URX34M_hist <- svyhist(~scale(log2(URX34M)), ewas_wt_subset, probability = F, breaks = 30,
                       las = 1, xlab = '', ylab = '', main = '3-/4-methylhippuric acid')

URXBP3_hist <- svyhist(~scale(log2(URXBP3)), ewas_wt_subset, probability = F, breaks = 30,
                       las = 1, xlab = '', ylab = '', main = 'Urinary benzophenone-3')

URXMZP_hist <- svyhist(~scale(log2(URXMZP)), ewas_wt_subset, probability = F, breaks = 30,
                       las = 1, xlab = '', ylab = '', main = 'Mono-benzyl phthalate')

URXUTU_hist <- svyhist(~scale(log2(URXUTU)), ewas_wt_subset, probability = F, breaks = 30,
                       las = 1, xlab = '', ylab = '', main = 'Urinary tungsten')


mtext('Chemical measurement, log'[2]~'transformed and standardized', side = 1, line = -1, outer = T)

mtext('Weighted number of measures', side = 2, line = 2, outer = T)

dev.off()


```

# Correlation matrix
## Data prep
```{r}

# load chem dictionary
use_these_chems <- read_csv('Data/use_these_chems.csv')

# select the chem names
chem_list <- use_these_chems %>%
  pull(variable_codename_use)

# Select the participants to use
seqn_included <- ewas_data %>% 
  filter(include == 1) %>% 
  pull(SEQN)

# select those chemicals and log transform values
corr_chems <- chems %>%
  mutate(URXUCD = case_when(URXUCD == 0 ~ NA,
                            .default = URXUCD)) %>%
  filter(SEQN %in% seqn_included) %>%
  select(all_of(chem_list)) %>%
  mutate(across(where(is.character), ~na_if(., "."))) %>%
  mutate_all(function(x) as.numeric(as.character(x))) %>%
  log2()

# Rename chemical family variable for the plot legend later
use_these_chems$`Chemical Family` <- use_these_chems$chemical_family

# Define a vector of chemical family names in a particular order
unique(use_these_chems$`Chemical Family`)
chem_family_levels <- c("Acrylamide", 
                        "Phosphate Flame Retardants (PFR)", 
                        "Polychlorinated Biphenyls (PCB)", 
                        "Dioxins", 
                        "Metals", 
                        "Phthalates & Plasticizers", 
                        "Personal Care & Consumer Product Compounds", 
                        "Pesticides", 
                        "Aromatic Amines", 
                        "Polyaromatic Hydrocarbons (PAH)", 
                        "Volatile Organic Compounds (VOC)", 
                        "Smoking Related Compounds", 
                        "Per- and Polyfluoroalkyl Substances (PFAS)", 
                        "Aldehydes")

# Ensure that the levels of the chemical family are in a defined order to ensure proper color scheme
use_these_chems$`Chemical Family` <- factor(use_these_chems$`Chemical Family`, 
                                            levels = chem_family_levels)

# make sure the dataset to color the chemicals is in the right order
chem_fam_reorder <- use_these_chems %>%
  arrange(`Chemical Family`)
chem_order <- chem_fam_reorder$variable_codename_use

# make the chemicals in the right order too
chem_cor_order <- corr_chems %>%
  select(all_of(chem_order))

```

## Calculate correlations
```{r}
chem_correlations <- cor(chem_cor_order,
                         use = "pairwise.complete.obs",
                         method = "spearman")
```

## Correlation table
```{r}
# Create a dataframe of the column names and chem names for the correlations
identical(colnames(chem_correlations), rownames(chem_correlations))

conversion_update <- use_these_chems %>%
  rename(chem_colnames = variable_codename_use) %>%
  select(chem_colnames, variable_description_use) %>%
  mutate(variable_description_use = gsub("\\s\\(([^()]+)\\)$", 
                                         "", 
                                         variable_description_use))

chem_colnames <- colnames(chem_correlations)
chem_colnames_df <- data.frame(chem_colnames)

# Attach the chemical names to the codenames in the order they appear in the correlation matrix
chem_names <- left_join(chem_colnames_df, 
                        conversion_update, 
                        by = "chem_colnames")

# Replace the column and row names of the chemical correlation matrix with the actual chemical names
chem_correlations_copy <- chem_correlations
colnames(chem_correlations_copy) <- chem_names$variable_description_use
rownames(chem_correlations_copy) <- chem_names$variable_description_use

# Save
write_csv(chem_correlations_copy, file = 'Output Cog ExWAS/chem_correlations_matrix.csv')

```

## Plot prep
```{r}
# set up to make the chemical family color bars
chem_fam_names <- chem_fam_reorder %>%
  select(`Chemical Family`) %>% 
  as.data.frame()

row.names(chem_fam_names) <- chem_fam_reorder$variable_codename_use

# identify where to put the gaps between chemical families
num_chem_fam <- chem_fam_names %>%
  group_by(`Chemical Family`) %>%
  summarise(number = n()) %>%
  ungroup()

# calculate where the breaks in the plot should go
num_chem_fam$breaks <- cumsum(num_chem_fam$number)

# assign the colors to chem families
chem_family_colors = list(`Chemical Family` = c(Acrylamide = "#8B0000",
                                                `Phosphate Flame Retardants (PFR)` = "#FF6B00",
                                                `Polychlorinated Biphenyls (PCB)` = "#FF69B4",
                                                `Dioxins` = "#FFA500",
                                                `Metals` = "#228B22",
                                                `Phthalates & Plasticizers` = "#A4D3EE",
                                                `Personal Care & Consumer Product Compounds` = "#A2CD5A",
                                                `Pesticides` = "#1E90FF",
                                                `Aromatic Amines` = "#be67c9",
                                                `Polyaromatic Hydrocarbons (PAH)` = "#cf9b76",
                                                `Volatile Organic Compounds (VOC)` = "#828282",
                                                `Smoking Related Compounds` = "#8B4513",
                                                `Per- and Polyfluoroalkyl Substances (PFAS)` = "#FFB6C1",
                                                `Aldehydes` = "#0E1171"))

# drop the units from the chemical names
chem_fam_reorder$variable_description_use <- gsub("\\s\\(([^()]+)\\)$", 
                                                  "", 
                                                  chem_fam_reorder$variable_description_use)

chem_fam_reorder <- chem_fam_reorder %>%
  droplevels("Chemical Family")

row.names(chem_fam_names) <- chem_names$variable_description_use

```

## Plot
```{r}

# https://stackoverflow.com/questions/31677923/set-0-point-for-pheatmap-in-r
paletteLength <- 100
myColor <- colorRampPalette(c("blue", "white", "yellow", "red"))(paletteLength)
# length(breaks) == length(paletteLength) + 1
# use floor and ceiling to deal with even/odd length pallettelengths
myBreaks <- c(seq(min(chem_correlations, na.rm = T), 0, length.out=ceiling(paletteLength/2.8) + 1),
              seq(max(chem_correlations, na.rm = T)/paletteLength,
                  max(chem_correlations, na.rm = T),
                  length.out=floor(paletteLength/1.8)))

pdf(file = "Output Cog ExWAS/chemical_heatmap_correlation_smk.pdf", 
    width = 14, height = 9)

pheatmap(chem_correlations_copy,
         cluster_rows = FALSE, cluster_cols = FALSE,
         annotation_row = chem_fam_names,
         annotation_col = chem_fam_names,
         labels_row = chem_fam_reorder$variable_description_use, # chemical names
         labels_col = chem_fam_reorder$variable_description_use,
         gaps_col = num_chem_fam$breaks,
         gaps_row = num_chem_fam$breaks,
         color = myColor, breaks = myBreaks,
         annotation_colors = chem_family_colors,
         annotation_names_row = FALSE, #x label
         annotation_names_col = FALSE, #y label
         fontsize_row = 4,
         fontsize_col = 4,
         legend = TRUE)

dev.off()
print("chemical_heatmap_correlation_smk.pdf")


svg(file = "Output Cog ExWAS/chemical_heatmap_correlation_smk.svg", 
    width = 14, height = 9)

pheatmap(chem_correlations,
         cluster_rows = FALSE, cluster_cols = FALSE,
         annotation_row = chem_fam_names,
         annotation_col = chem_fam_names,
         labels_row = chem_fam_reorder$chemical_name, # chemical names
         labels_col = chem_fam_reorder$chemical_name,
         gaps_col = num_chem_fam$breaks,
         gaps_row = num_chem_fam$breaks,
         color=myColor, breaks=myBreaks,
         annotation_colors = chem_family_colors,
         annotation_names_row = FALSE, #x label
         annotation_names_col = FALSE, #y label
         fontsize_row = 2,
         fontsize_col = 2,
         legend = TRUE)

dev.off()

```
